Removed circular import
